Pulling in upstream

+++ controller/lib/src/entity_descriptor_imp.cpp
@@ -35,7 +35,7 @@
 
 namespace avdecc_lib
 {
-    entity_descriptor_imp::entity_descriptor_imp(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len) : descriptor_base_imp(end_station_obj)
+    entity_descriptor_imp::entity_descriptor_imp(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len) : descriptor_base_imp(end_station_obj)
     {
         desc_entity_read_returned = jdksavdecc_descriptor_entity_read(&entity_desc, frame, pos, frame_len);
 
@@ -162,12 +162,12 @@ namespace avdecc_lib
         return entity_desc.current_configuration;
     }
 
-    void entity_descriptor_imp::store_config_desc(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len)
+    void entity_descriptor_imp::store_config_desc(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len)
     {
         config_desc_vec.push_back(new configuration_descriptor_imp(end_station_obj, frame, pos, frame_len));
     }
 
-    uint32_t STDCALL entity_descriptor_imp::config_desc_count()
+    size_t STDCALL entity_descriptor_imp::config_desc_count()
     {
         return config_desc_vec.size();
     }
@@ -213,7 +213,7 @@ namespace avdecc_lib
         return default_send_acquire_entity_cmd(this, notification_id, acquire_entity_flag);
     }
 
-    int entity_descriptor_imp::proc_acquire_entity_resp(void *&notification_id, const uint8_t *frame, uint16_t frame_len, int &status)
+    int entity_descriptor_imp::proc_acquire_entity_resp(void *&notification_id, const uint8_t *frame, size_t frame_len, int &status)
     {
         return default_proc_acquire_entity_resp(aem_cmd_acquire_entity_resp, notification_id, frame, frame_len, status);
     }
@@ -223,7 +223,7 @@ namespace avdecc_lib
         return default_send_lock_entity_cmd(this, notification_id, lock_entity_flag);
     }
 
-    int entity_descriptor_imp::proc_lock_entity_resp(void *&notification_id, const uint8_t *frame, uint16_t frame_len, int &status)
+    int entity_descriptor_imp::proc_lock_entity_resp(void *&notification_id, const uint8_t *frame, size_t frame_len, int &status)
     {
         return default_proc_lock_entity_resp(aem_cmd_lock_entity_resp, notification_id, frame, frame_len, status);
     }
